requirements.txt
Django==4.2
python-dotenv==1.0.1
tortoise-orm==0.21.7